+++ crates/cargo_utils/src/lib.rs
@@ -3,6 +3,7 @@ mod fs_utils;
 mod local_manifest;
 mod manifest;
 mod registry;
+mod token;
 mod version;
 mod workspace_members;
 
@@ -11,6 +12,7 @@ pub use fs_utils::*;
 pub use local_manifest::*;
 pub use manifest::*;
 pub use registry::*;
+pub use token::*;
 pub use version::*;
 pub use workspace_members::*;
